feat: support basic component mode (#67)
* feat: support basic component mode

* chore: lint

* fix: rename config
