+++ examples/react-component-with-mulit-demo/build.json
@@ -1,4 +1,5 @@
 {
+  "basicRepo": true,
   "plugins": [
     "build-plugin-component",
     "build-plugin-fusion",
